+++ src/dotnet.nugit/ApplicationService.cs
@@ -30,7 +30,7 @@ namespace dotnet.nugit
             this.restorePackagesCommand = restorePackagesCommand ?? throw new ArgumentNullException(nameof(restorePackagesCommand));
 
             this.app.Help(Resources.Resources.Program_Help_description);
-            
+
             this.Initialize();
         }