+++ src/dotnet.nugit/Services/DotnetUtility.cs
@@ -1,15 +1,20 @@
 ﻿namespace dotnet.nugit.Services
 {
     using System;
+    using System.Collections.Generic;
+    using System.Collections.Immutable;
+    using System.ComponentModel.DataAnnotations;
     using System.Diagnostics;
     using System.IO;
     using System.IO.Abstractions;
+    using System.Linq;
     using System.Text;
     using System.Threading;
     using System.Threading.Tasks;
+    using Abstractions;
     using Microsoft.Extensions.Logging;
-    using nugit.Abstractions;
     using Resources;
+    using Workspace;
 
     internal sealed class DotNetUtility(
         IFileSystem fileSystem,
@@ -18,29 +23,57 @@
         private readonly IFileSystem fileSystem = fileSystem ?? throw new ArgumentNullException(nameof(fileSystem));
         private readonly ILogger<DotNetUtility> logger = logger ?? throw new ArgumentNullException(nameof(logger));
 
-        public async Task BuildAsync(string projectFile, TimeSpan? timeout = null, CancellationToken cancellationToken = default)
+        public async Task BuildAsync(IDotNetProject project, TimeSpan? timeout = null, CancellationToken cancellationToken = default)
         {
-            if (string.IsNullOrWhiteSpace(projectFile)) throw new ArgumentException(Resources.ArgumentException_Value_cannot_be_null_or_whitespace, nameof(projectFile));
+            ArgumentNullException.ThrowIfNull(project);
 
+            string? projectFile = project.GetProjectFilePath();
             string? workingDirectoryPath = this.fileSystem.Path.GetDirectoryName(projectFile);
 
             var arguments = $"build \"{projectFile}\" --configuration Release";
             await this.RunDotNetProcessAsync(arguments, workingDirectoryPath, timeout, cancellationToken);
         }
 
-        public async Task<bool> TryPackAsync(string projectFile, string packageTargetFolderPath, PackOptions options, TimeSpan? timeout = null, CancellationToken cancellationToken = default)
+        public async Task<bool> TryPackAsync(IDotNetProject project, string packageTargetFolderPath, PackOptions options, TimeSpan? timeout = null, CancellationToken cancellationToken = default)
         {
-            if (string.IsNullOrWhiteSpace(projectFile)) throw new ArgumentException(Resources.ArgumentException_Value_cannot_be_null_or_whitespace, nameof(projectFile));
+            ArgumentNullException.ThrowIfNull(project);
             if (string.IsNullOrWhiteSpace(packageTargetFolderPath)) throw new ArgumentException(Resources.ArgumentException_Value_cannot_be_null_or_whitespace, nameof(packageTargetFolderPath));
 
+            string? projectFile = project.GetProjectFilePath();
             string? workingDirectoryPath = Path.GetDirectoryName(projectFile);
 
-            // TODO: load the project, reflect all NuGet-specific properties
-            // Emit a (temporary) .nuspec file and build the package using: dotnet pack ~/projects/app1/project.csproj -p:NuspecFile=~/projects/app1/project.nuspec -p:NuspecBasePath=~/projects/app1/nuget
+            IProjectPackageMetadata packageSpec = project.DerivePackageSpec();
+            packageSpec.AddDefaultValues(ProjectPackageMetadata.Authors, $"{Environment.UserName}@{Environment.MachineName}");
+            if (packageSpec.Validate(out ICollection<ValidationResult> validationResults) == false)
+            {
+                var errors = new StringBuilder();
+                ImmutableList<string?> errorMessages = validationResults.Select(result => result.ErrorMessage).Where(s => string.IsNullOrWhiteSpace(s) == false).ToImmutableList();
+                errors.AppendJoin(Environment.NewLine, errorMessages);
+                this.logger.LogError("Failed to derive valid NuGet package specification.");
+                this.logger.LogError(errors.ToString());
+                return false;
+            }
 
-            var arguments = $"pack \"{projectFile}\" --output \"{packageTargetFolderPath}\" --version-suffix \"{options.VersionSuffix}\"";
+            string? nuspecFilePath = project.GetNuspecFilePath();
+            if (string.IsNullOrWhiteSpace(nuspecFilePath))
+                return false;
+
+            await this.WriteNuspecFile(packageSpec, nuspecFilePath, cancellationToken);
+
+            string? nuspecBasePath = project.GetNuspecBasePath();
+
+            // dotnet pack ~/projects/app1/project.csproj -p:NuspecFile=~/projects/app1/project.nuspec -p:NuspecBasePath=~/projects/app1/nuget
+            var arguments = $"pack \"{projectFile}\" -p:NuspecFile=\"{nuspecFilePath}\" -p:NuspecBasePath=\"{nuspecBasePath}\" --output \"{packageTargetFolderPath}\" --version-suffix \"{options.VersionSuffix}\"";
             int exitCode = await this.RunDotNetProcessAsync(arguments, workingDirectoryPath, timeout, cancellationToken);
+
             return exitCode == 0;
+        }
+
+        private async Task WriteNuspecFile(IProjectPackageMetadata packageSpec, string nuspecFilePath, CancellationToken cancellationToken)
+        {
+            await using Stream stream = this.fileSystem.File.Open(nuspecFilePath, FileMode.OpenOrCreate, FileAccess.Write, FileShare.None);
+            stream.SetLength(0);
+            await packageSpec.WriteNuspecFileAsync(stream, cancellationToken);
         }
 
         private async Task<int> RunDotNetProcessAsync(string arguments, string? workingDirectory, TimeSpan? timeout = null, CancellationToken cancellationToken = default)