+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- [PR #8] The application can access package configuration properties directly from project files and override/tweak them as needed. The `add` and `restore` commands now support the `--configuration` flag, as known from other `dotnet` commands. (#7)
+- Adds support for F# project files (#7)
+
+### Changed
+
+- [PR #8] Changes services involved with building and packing NuGet packages.
+- [PR #8] The `dotnet pack` command is now invoked with the `-p:NuspecFile` option; Nuspec files are dynamically generated from project files (#7)
+
+
+### Fixed
+
+- [PR #8] Fixes broken task factories
+
 ## [0.2.0-alpha1.240419.1] - 2024-04-19
 
 ### Added